refactor: delegate racing functionality to racing class.
