+++ src/main/java/racingcar/controller/RaceController.java
@@ -1,7 +1,9 @@
 package racingcar.controller;
 
+import java.util.List;
 import racingcar.model.car.Cars;
-import racingcar.model.judge.WinnerFinder;
+import racingcar.model.race.Racing;
+import racingcar.model.race.Winner;
 import racingcar.view.InputView;
 import racingcar.view.OutputView;
 
@@ -15,24 +17,23 @@ public class RaceController {
     }
 
     public void execute() {
-        final Cars cars = createCars();
+        final Cars cars = new Cars(carNames);
+        final Racing racing = new Racing(cars);
         OutputView.printResultMessage();
-
         for (int i = 0; i < tryCount; i++) {
-            cars.startRacing();
-            OutputView.printResult(cars);
+            runSingleTrial(cars, racing);
         }
-
-        printWinner(cars);
+        OutputView.printWinner(findWinner(cars));
     }
 
-    private Cars createCars() {
-        return new Cars(carNames);
+    private void runSingleTrial(final Cars cars, final Racing racing) {
+        racing.startRacing();
+        OutputView.printResult(cars);
     }
 
-    private void printWinner(final Cars cars) {
-        final WinnerFinder winnerFinder = new WinnerFinder(cars);
-        OutputView.printWinner(winnerFinder.find());
+    private List<String> findWinner(final Cars cars) {
+        final Winner winner = new Winner(cars);
+        return winner.find();
     }
 
     private String inputCarNames() {
